+++ src/index.py
@@ -4,6 +4,8 @@ import db_helper
 
 
 def main():
+    """Alustaa tietokannan ja käyttöliittymän
+    """
     db_helper.drop_tables()
     db_helper.create_db()
     root = Tk()