+++ src/tests/services/user_service_test.py
@@ -11,36 +11,34 @@ class TestUserService(unittest.TestCase):
         self.password = "password"
         db_helper.drop_tables()
         db_helper.create_db()
+        self.service = UserService()
+
+    def tearDown(self):
+        self.service.conn.close()
 
     def test_registration_succeeds(self):
-        service = UserService()
-        self.assertTrue(service.register(
+        self.assertTrue(self.service.register(
             self.username, self.password, "password"))
 
     def test_registration_fails_with_short_password(self):
-        service = UserService()
         self.assertRaises(InvalidPasswordError,
-                          lambda: service.register("username", "p", "p"))
+                          lambda: self.service.register("username", "p", "p"))
 
     def test_registration_fails_if_username_exists(self):
-        service = UserService()
-        service.register(self.username, self.password, self.password)
-        self.assertRaises(UsernameExistsError, lambda: service.register(
+        self.service.register(self.username, self.password, self.password)
+        self.assertRaises(UsernameExistsError, lambda: self.service.register(
             self.username, "anotherpassword", "anotherpassword"))
 
     def test_registration_fails_if_passwords_dont_match(self):
-        service = UserService()
-        self.assertRaises(InvalidPasswordError, lambda: service.register(
+        self.assertRaises(InvalidPasswordError, lambda: self.service.register(
             self.username, "password", "anotherpassword"))
 
     def test_login_succeeds(self):
-        service = UserService()
-        service.register(self.username, self.password, "password")
-        self.assertTrue(service.login(self.username, self.password))
-        self.assertEqual(service.current_user.username, "username")
+        self.service.register(self.username, self.password, "password")
+        self.assertTrue(self.service.login(self.username, self.password))
+        self.assertEqual(self.service.current_user.username, "username")
 
     def test_login_fails_with_invalid_credentials(self):
-        service = UserService()
-        service.register(self.username, self.password, "password")
+        self.service.register(self.username, self.password, "password")
         self.assertRaises(InvalidCredentialsError,
-                          lambda: service.login(self.username, "wrong"))
+                          lambda: self.service.login(self.username, "wrong"))