+++ src/ui/ui.py
@@ -8,11 +8,26 @@ from os import getenv
 
 
 class UI:
+    """Luokka, joka hallinnoi sovelluksen näkymiä
+
+    Attributes:
+        root: juuri
+        current_view: näkymä, jossa ollaan
+    """
     def __init__(self, root):
+        """
+        Alustaa UI-luokan ja asettaa juurielementin sekä nykyisen näkymän.
+
+        Args:
+            root: juuri
+        """
         self.root = root
         self.current_view = None
 
     def start(self):
+        """Näyttää aloitusnäkymän
+        """
+
         if self.current_view:
             self.current_view.destroy()
 
@@ -24,18 +39,27 @@ class UI:
             self.current_view.initialize_view()
 
     def show_register_view(self):
+        """Näyttää rekisteröitymisnäkymän
+        """
+
         if self.current_view:
             self.current_view.destroy()
         self.current_view = RegisterView(self.root, self.show_index_view)
         self.current_view.initialize_view()
 
     def show_login_view(self):
+        """Näyttää kirjautumisnäkymän
+        """
+
         if self.current_view:
             self.current_view.destroy()
         self.current_view = LoginView(self.root, self.show_index_view)
         self.current_view.initialize_view()
 
     def show_index_view(self):
+        """Näyttää listan kursseista
+        """
+
         if self.current_view:
             self.current_view.destroy()
         self.current_view = IndexView(
@@ -43,13 +67,21 @@ class UI:
         self.current_view.initialize_view()
 
     def show_create_course_view(self):
+        """Näyttää kurssien luomisnäkymän
+        """
+
         if self.current_view:
             self.current_view.destroy()
         self.current_view = CreateCourseView(self.root, self.show_index_view)
         self.current_view.initialize_view()
 
     def show_course_view(self, course):
+        """Näyttää yksittäisen kurssin tiedot
+
+        Args:
+            course: kurssi, jonka tiedot näytetään
+        """
         if self.current_view:
             self.current_view.destroy()
         self.current_view = CourseView(self.root, self.show_index_view)
-        self.current_view.initialize_view(course)
+        self.current_view.initialize_view(course.id)